day20 미션

+++ section12-2/src/components/DiaryItem.jsx
@@ -6,11 +6,16 @@ import { DiaryDispatchContext } from "../App";
 import { useContext } from "react";
 
 const DiaryItem = ({ id, emotionId, createdDate, content }) => {
+  // DiaryDispatchContext -> onDelete 받아오기
   const { onDelete } = useContext(DiaryDispatchContext);
   const nav = useNavigate();
 
+  // 삭제시 실행할 이벤트 함수
   const onClickDelete = () => {
-    onDelete(id);
+    if (window.confirm("정말 삭제합니까?")) {
+      onDelete(id);
+      alert("삭제되었습니다.");
+    }
   };
 
   return (
@@ -31,7 +36,8 @@ const DiaryItem = ({ id, emotionId, createdDate, content }) => {
 
       <div className="button_section">
         <Button onClick={() => nav(`/edit/${id}`)} text={"수정하기"} />
-        <Button type={"NEGATIVE"} onClick={onDelete} text={"삭제하기"} />
+        {/* 삭제 버튼 추가 */}
+        <Button type={"NEGATIVE"} onClick={onClickDelete} text={"삭제하기"} />
       </div>
     </div>
   );
